Handle logs with only reliable messages from a boot
We have logs with only a reliable message from the previous boot, and
both reliable and unreliable messages from the subsequent boot.  These
are still sortable, though the logic needed to be expanded.

There are still small holes here with enough random combinations and
boots, but let's tackle those when they happen for real.

Change-Id: Ie9eabf2662216c8fd30ea764e6f78039e7eba60e
Signed-off-by: Austin Schuh <austin.linux@gmail.com>