Fix AOS support for realtime replay
This patch fixes the use-case where you provide a single event loop and
just want to replay logged events from the perspective of that event
loop's node.

Includes a test running a ShmEventLoop against a single-node logfile,
and running a multi-node replay into a single EventLoop in
simulation (the choice of single vs multi node here is arbitrary--it
should work with both single and multi node configs in both
simulation and shm).

This has a few caveats:
* Doesn't replay remote timestamps currently.
* Doesn't correct for implied changes in node<->node offsets due to
  changes in the clock.
* Had to add a flag to choose how to manage fetcher behavior for
  messages from before the start of the log.

Change-Id: I8f101e8774e0923bacc4f7e1bf58c9da02fd0d3f
Signed-off-by: James Kuszmaul <james.kuszmaul@bluerivertech.com>