Add a cursor to graph in Spline UI
This connects the graph to the robot that gets drawn along the spline.
As you move your mouse along the spline, you can see what part of the
graph it affects, and when you move your mouse along the graph, the
robot traces out the path of the spline.

This change lays the groundwork for being able to apply velocity limits
at specific points along the spline in the future.

Change-Id: I7d850f27fb25104c0b9d678b6701d5d74a11b05f
Signed-off-by: Ravago Jones <ravagojones@gmail.com>