Require non-negative times for AOS timers
Linux doesn't like negative monotonic clock times, so disallow them.

In order to support this, LogReader now does not schedule events at
times known to be in the past (this should be a no-op for simulated
event loops, but for realtime replay with non-zero clock offsets will
avoid scheduling events too far in the past).

References: PRO-14846
Change-Id: Ieb4bf58c5e1083ff30c7df7bfddff356c95ba9ba
Signed-off-by: James Kuszmaul <james.kuszmaul@bluerivertech.com>