Deduplicate timestamps
Mostly in tests, but maybe possible in the real world, we sometimes get
two messages arriving at the same time with the same offset.  In this
case, we add them both to the filter, and then try to interpolate.  This
results in a 0/0 calculation for that time, which goes quite poorly.
Deduplicate those and test.

Change-Id: Ide7de1a00c85d75cc6253a9ca1ec30100e5d9181
Signed-off-by: Austin Schuh <austin.linux@gmail.com>