Add a timestamps_empty() call to simplify multiple filters
Most of the times we want to know timestamp size, we are really looking
to see if it's empty or not.  With multiple boots, empty is an easier
construct than size to compute, and just as useful.  Add a way to do
that instead.  Size is mostly retained for tests because it is useful
there.

Change-Id: Ia71255e642d12b6557287df6d8f81c1d5ca99e9b
Signed-off-by: Austin Schuh <austin.schuh@bluerivertech.com>