Add a pointer for caching to NoncausalTimestampFilter
This makes repeated lookups a lot cheaper since we can save which
timestamps we used last time, and can reuse them if the solution is
similar.  This gives us a spot in the future to do a more clever binary
search.  This all assumes that subsequent lookups are likely to be for
similar times.

Change-Id: Ic137a16eb98fa5b0a63fc140bd20aae855eaef8b
Signed-off-by: Austin Schuh <austin.schuh@bluerivertech.com>