Handle large, single direction timestamp outages when estimating time
Brian Smartt had the clever idea of taking the max of both directions.
This essentially warps any long stretches of time such that they always
stay above the other direction with more fresh timestamps.  Because the
lines then don't cross, we can solve like we used to.

We only really need to introduce this concept into FindTimestamps.
Everything else works off a valid timestamp pair so it stays the same.

Add a test to reproduce this behavior (it happens occasionally in the
wild) and then fix it.

Change-Id: I999280fc1db1f178835092a8d024a9862144d7cb
Signed-off-by: Austin Schuh <austin.schuh@bluerivertech.com>