Remove a n^2 operation in timestamp_extractor
The multinode timestamp estimator holds a deque of all the solved
timestamps, and removes them when enough time elapses.  We were never
teaching the estimator that time had elapsed, so it was never expiring
entries.  There is a linear step processing all old solved timestamps
looking for what we've written already, which was ending up as a n^2
operation.  Drop the old timestamps like the real system...

This took a sort of 10 nodes with 10 GB of compressed events down from
days to complete to around an hour.

Change-Id: I126139ee3e572cd535127d50875bb80fd000dbdf
Signed-off-by: Austin Schuh <austin.linux@gmail.com>