Add support for multithreaded lzma encoder
liblzma already has support for a multithreaded encoder.  The plumbing
needed to enable it is pretty small, and can be done at runtime.  This
sets us up to compress more than 1 core of data, potentially.

I did some basic profiling testing, and the block size should be left at
the default for optimal performance.  I made it configurable otherwise
it is so big that the tests don't pass.  Decreasing the block size
increases overhead and decreases performance significantly.  The timeout
parameter seems to have little effect for reasonable workloads.

Change-Id: Ia886bc8576db19f1b59a76b41dc7cdc2e85d52bd
Signed-off-by: Austin Schuh <austin.schuh@bluerivertech.com>