Compute the gradient and hessian at the same time
We are already iterating over everything.  It's simpler to save out both
results at once and return them in a structure.

We want to add constraints so we get a constrained optimization
problem.  Since the goal is essentially to keep the error > 0, we can
use an active set method based off each computed error.  That is going
to be most efficient to do if we compute which are the active
constraints as we solve.  Hence refactoring to provide a single function
to do it all at once.

Change-Id: Iadfe672743db8c8856299ecab05c1844f859d94b
Signed-off-by: Austin Schuh <austin.schuh@bluerivertech.com>