Use 128 bit math for ExtrapolateOffset
We were being lazy with ExtrapolateOffset and using the point as the
integer part of the result, and putting all of the error in the double.
That gets imprecise when we extrapolate long distances, which is common
at the start of log files.

Instead, use the exact integer math, and use the double to capture the
remainder.  This is significantly more precise, since we should never
see values over 1.0 in the double.

Change-Id: If8e2029a941494734f239bf43220ea698c9b998b
Signed-off-by: Austin Schuh <austin.linux@gmail.com>