Add a worst case bound to NoncausalTimestampFilter
With a bounded slew rate, we can define a function which is the worst
that time could change.  Since the worst time could change is
represented by the extrapolation function, this is the max of the
extrapolation function for both bounding points.

This lets us start to enforce constraints on the worst case, rather
than on the average.  Given what we've seen when we have outages, I
think that will close up a hole in our current logic.  Usage will follow
soon.

Now that we are good at testing 128 bit math, just do it with 128 bit
math from the beginning.

Change-Id: I1dcb1522c39468dd742b8f97d77753fa57737c70
Signed-off-by: Austin Schuh <austin.linux@gmail.com>