Use 128 bit math for InterpolateOffset
We were being lazy with InterpolateOffset and using the first point as
the integer part of the result, and putting all of the error in the
double.  That gets imprecise when we extrapolate long distances, which
is common at the start of log files.

Instead, use the exact integer math, and use the double to capture the
remainder.  This is significantly more precise, since we should never
see values over 1.0 in the double.

Change-Id: I02bf511239cd20dbff7d17327db2f665c6aa41d3
Signed-off-by: Austin Schuh <austin.linux@gmail.com>