Improve precision of the gradient calculation
We were computing the integer and double errors, and then adding those
and converting to a double.  At the start, when the bounds is far apart,
a lot of digits end up being consumed by capturing the distance to those
two constraints, which are then subtracted out.  This reduces precision.

Instead, accumulate up the integer portions separately and only combine
them after finishing.

Change-Id: I739eb370786860022d653e918e7538deee4b9699
Signed-off-by: Austin Schuh <austin.schuh@bluerivertech.com>