Stop popping when finding invalid solutions
When 2 solutions in the middle of a solve are invalid, we were popping
one.  This can cause some serious issues.  Imagine the following case:

We've got 3 nodes to solve for.  Their clocks are all synced for sake of
explination.

ta = 5.0
tb = 5.0
tc = 1.0

We really want to solve for tc, but we might learn when solving that ta
and tb are invalid but close enough to ignore.  So, before, we would
have popped tc, freezing things past the final ta solution.  Instead, we
should just wait to deal with ta and tb until one of them is the best
solution and has been popped.

Change-Id: Ib7b3682c9b385bfbb6bc2f173647915f2affbdf0
Signed-off-by: Austin Schuh <austin.schuh@bluerivertech.com>