Use a Primal-Dual Interior-Point method solver
See https://web.stanford.edu/~boyd/cvxbook/bv_cvxbook.pdf#page=623 for
more background.

In certain cases, the unconstrained solver will converge to a solution
which is outside the feasible region.  This happens when there is a loop
in node connectivity graph, and all but one of the legs of the loop are
high confidence (lots of packets flowing), and the last leg is low
confidence and pulls the mean away.

To solve this, we really want to treat the bounds lines as both part of
the cost, and constraints.  We'd like something that is "centered" for
some definition of centered, but really need something which satisfies
the constraints.

Like the original unconstrained solver, this solver needs to be special.
It has the same precision requirements, and also has to have special
logic to handle the discontinuity across the vertices in the bounds
lines.

The constrained solver is much much slower than the unconstrained
solver.  To accommodate this, first try solving unconstrained as a seed,
and, once the unconstrained solve converges, then see if we need to
solve constrained.  This also gives us a decent seed to start with.

As currently implemented, the constrained solver doesn't seem to want
to solve the simultaneous problem, so we fall back to the sequential
problem.  This makes it even slower.

Next steps are:
* Figure out why we need so many iterations to converge.  Can we find a
  better seed?
* Solve the simultaneous problem.
* Can we estimate which constraints will be relevant at the start and
  remove some of them?
* The slack variable can be added as part of solving instead of adding
  it explicitly.  This should save some allocations.

Change-Id: Ifc9cdf3aa2c3819931f2d7420aa4d2908878d78a
Signed-off-by: Austin Schuh <austin.schuh@bluerivertech.com>