Fix constrained simultaneous solver
Turns out we weren't passing through the solution node so it was popping
the wrong thing.  This is like a 6x speedup for the test logs I've got
since we don't solve 10 problems anymore.

To find this, I had to clean up debugging quite a bit to be able to see
and understand the problem.  Things should be more clear now.

Change-Id: I7525de16f419fc379aaf841214bc0ec90b721e62
Signed-off-by: Austin Schuh <austin.schuh@bluerivertech.com>