Only solve with the constraints which are relevant
With a large problem statement, typically only a couple of constraints
are relevant in the end.  Turns out, the set of relevant constraints is
highly related to the set of constraints which are initially violated by
the unconstrained solution.  So, guess that those are the relevant set,
and resolve with any constraints which end up being violated.  Resolving
occasionally is cheaper than solving the original problem.

This reduces the solve time by another 3x.  Should be about as good as
we can do without guessing a better seed.

Change-Id: I4f73e6b9a702b38c2bac2f9a258162f049279fd6
Signed-off-by: Austin Schuh <austin.schuh@bluerivertech.com>