Make aos_config rule outputs predeclared
This makes it easy to have rules depend on a specific one of the outputs
(i.e., just one of the .json, .stripped.json, or .bfbs) without having
to write an entire bazel rule yourself.

It also means that we now actually catch it when someone accidentally
names the input config the same as the resulting generated config.

Change-Id: Id36388f52794aeaef4e3b9357cb495f88ac5f313
Signed-off-by: James Kuszmaul <jabukuszmaul+collab@gmail.com>