Refactor TimestampProblem into a solver and a problem statement
This was a great suggestion from Quan and Brian Smartt.  This lets me
write tests easier by injecting example problems.  The solver is still
very much geared towards the timestamp problem, but it has a much better
hope of working for other problems.

It is hard enough to separate all this now that I'm not separating this
out and distributing it into the existing patches and commiting it last.

Change-Id: I685fb4e6642314db7e36c6e8f7440facf9078a62
Signed-off-by: Austin Schuh <austin.schuh@bluerivertech.com>