Migrate everything to upstream pip packages
This patch migrates all the Python packages I could find to the
packages provided by rules_python. It's possible that there are some
that I missed. In that case, the finder will have to migrate it on
their own.

A future patch will remove the remaining references to the
Debian-packaged Python packages in future patches. This patch here is
solely focused on migrating the existing users.

Signed-off-by: Philipp Schrader <philipp.schrader@gmail.com>
Change-Id: Ic7a8ba6110516aa2446ad2ad4e2495fab62d824c