Import GTK in preparation for rules_python
We currently import GTK libraries as part of the "python_gtk" repo.
That pulls in GTK, but also Python and a bunch of pip packages.

This patch makes it so that GTK is pulled in via a dedicated repo.
That allows us to have dedicated repos for pip packages via
rules_python.

Future patches will make use of the new GTK import.

I had to modify `download_packages.py` in order to ensure that `libc6`
really got excluded from the final tarball. On James' personal
machines, he observed stack smashing crashes when `libc` is in the
tarball. The `force_includes` usage pulled in `libc6` even though we
don't want it to.

Signed-off-by: Philipp Schrader <philipp.schrader@gmail.com>
Change-Id: Ia3ecf51e4bd4c7aa0e72646e4c2af738f6086c01