Found a workaround for the camera exposure issues we've seen
Calling SetExposure before the calls to configure the camera (image height,
width, frame rate) allows the camera exposure to come up properly.

If this isn't done, then the camera comes up with images that are darker
than the desired exposure, and then any following calls to camera_reader
have a brighter (proper) exposure.

NOTE: this changes the exposure values that we use, so we need to be sure
to verify that the exposure settings are still correct for indoor/outdoor
use in competition.

I think this also makes auto exposure come up properly-- we were having
some issues with this before.

Change-Id: Ia01913cd6e639565faad4a26ec1234649dc288ad
Signed-off-by: Jim Ostrowski <yimmy13@gmail.com>