Add mkdocs bazel rules
I want to add some more detailed AOS documentation. mkdocs seems like a
reasonable choice, since it allows markdown (so GitHub and Gerrit will
still be able to render it even for people not using mkdocs), and it
was reasonably straightforwards to integrate.

Now that I am calling some python code that is using pip in the exec
configuration, also needed to set --host_platform for the
k8_upstream_python config.

Change-Id: I987f531759170272686f6488271be732827d0f9b
Signed-off-by: James Kuszmaul <james.kuszmaul@bluerivertech.com>