Make logfile_utils_test use QueueSpan
The tests don't need to use all the complicated writing methods which
are designed for efficiency, not ease of use.  Instead, throw a subclass
inside logfile_utils_test and put an easy WriteMessage type method
inside it and use that in the tests.  That then converts to WriteSpan,
which may not be as performant, but will be easy to use.

Change-Id: Id911cef38645c3307dc5c6854853502ea224a739
Signed-off-by: Austin Schuh <austin.linux@gmail.com>