Add IsDisabled() to shm and simulated timers
This is an important addition especially for checking and testing if the
timers are disabled at the right place. Earlier, we were just assuming
that timers would be disabled when we call disable() or let the timer
expire, but this is just a cherry on top to check the same.

Change-Id: I8b47c48609a085efb83eda99236d60f5c1e8f3aa
Signed-off-by: James Kuszmaul <james.kuszmaul@bluerivertech.com>