Import missing pip packages for --config=k8_upstream_python
There are still a bunch of packages that were not available for
`rules_python`. This patch imports them.

I was running into some trouble with pip insisting on listing
`pyyaml_env_tag` as `pyyaml-env-tag` in the lock file. This caused
some trouble in the interaction between our wheel override and
rules_python. I adjusted both to use the same name.

Signed-off-by: Philipp Schrader <philipp.schrader@gmail.com>
Change-Id: Ibc501461deae41c4285f9149dd020678a2ee91b0