Scouting App: Prevent double tap zoom on counter buttons
In Rapid React, teams often shoot two cargo in rapid succession so the user of the scouting
app taps the increment button twice. However, this causes the app to zoom in on the button on
mobile devices. This causes the user to stop watching the match to undo the zoom-in and
a lot of frustration. More info on the solution can be viewed at
https://developer.mozilla.org/en-US/docs/Web/CSS/touch-action#manipulation


Change-Id: I5a86315bb521d401e5b5ca5959e3dd20ab330d0c
Signed-off-by: Filip Kujawa <filip.j.kujawa@gmail.com>