LogReplayer to replay log with selected channels on active nodes
Log replayer checks if the channels specified in the replayer json
config are available to replay. Remaps the channels which are needed to
be remapped. Plots the timing report to compare the replay times with
actual send times in the log and stdouts the specified config or the
logged config depending on the flags specified. And replays the messages
using a shared memory event loop to keep it realtime.

Change-Id: I9afbd5c2dfd309b544baffacfbc10bc5892943fd
Signed-off-by: James Kuszmaul <james.kuszmaul@bluerivertech.com>