Update all Python packages
This patch is the result of:

    $ bazel run //tools/python:requirements.update -- --upgrade
    $ bazel run //tools/python:mirror_pip_packages --config=k8_upstream_python -- --ssh_host software

I added special logic to exclude `pygobject` from being fixed with
`auditwheel`. For some reason `auditwheel` fixing causes issues at
runtime. I suspect it has to do with assumptions that the code makes
about files living relative to the path of the .so file, but I'm not
sure.

Either way, this pulls in the latest matplotlib. In particular, the
fix of interest is
https://github.com/matplotlib/matplotlib/issues/23074.
Unfortunately, this means that the `rules_python` toolchain does not
allow for tkagg plots. We'll probably have to use the Gtk runtime as a
work around.

Signed-off-by: Philipp Schrader <philipp.schrader@gmail.com>
Change-Id: I083dc0fcae8315a7a40a351abad3ae2df0031c5d