Fix race in ShmEventLoop
The logic in ShmEventLoop is trying to make sure messages are replayed
in order, and wakeups aren't dropped.  It does this roughly by:

* Grab the time
* Process all the wakeups
* Look at all the queues for new messages
* Process messages (and timers) up until the time grabbed above
* Repeat until there is nothing to do
* Go back to sleep

There is a small race.  If a message is published after the time is
grabbed but before wakeups are processed, we won't process it and will
go to sleep.  The fix is to retry in that case.

Change-Id: Ic4a64b6abf376192ebe415fe3c83bd7edc5889db
Signed-off-by: James Kuszmaul <james.kuszmaul@bluerivertech.com>