Check validity of sender/fetcher before operating on it
Now if a user tries to query the channel or context from a sender or
fetcher that isn't valid, we will fail with a clearer error instead of
segfaulting. Note that this can happen when the user creates the sender
or fetcher using the TryMake* API and operates on the result without
explicitly checking its validity first.

Change-Id: Ie30fe3c21b602346c16561d6425d8e0f4186ec9f
Signed-off-by: James Kuszmaul <james.kuszmaul@bluerivertech.com>