Update rules_python
This patch bumps rules_python from 0.12.0 to 0.16.1.

Our rules_python patch that allows us to override packages needed to
be rebased. The biggest upstream changes that happened upstream since
the last update is the addition of bzlmod support. This made the
rules_python patch look quite a bit different, but is functionally the
same.

Signed-off-by: Philipp Schrader <philipp.schrader@gmail.com>
Change-Id: I4db108d955e2fd0e63d69ea8f82e6bf5d0c1b6d8