Add missing nullptr check in LogReader::State constructor
This change only affects the code which uses replay filters.
We ran into the path where construction of State happens with no
timestamp_mapper while running log_replayer on single node logs.
The timestamp_mapper does not exist when there are no communications
with other nodes as no filters are needed for logfile sorting. This has been
reproduced by creating a unit test with a single node which has been blocked
of any communication with other nodes and by calling Register on the logged files.

Change-Id: Ie3e2f1742e0480339f572e474f5cb49a4f051525
Signed-off-by: James Kuszmaul <james.kuszmaul@bluerivertech.com>