Verify that the Cargo lockfile is complete
cargo-raze will happily just use the newest versions of anything missing
from here, which leads to CI failures when new versions are released.

Apparently we had a lot of dependencies that didn't meet the correct
version requirements, so they had to get upgraded along with this change
to validate those. I did verify that this isn't just upgrading
everything to the newest versions by doing that manually, it picks up a
few more.

Change-Id: I2c1363a977a159e54c376d50bf376a98fddde10c
Signed-off-by: Brian Silverman <bsilver16384@gmail.com>