Switch to new ftrace path
Tracing is typically mounted at /sys/kernel/tracing, and there's an
automounter in the kernel which mounts it on /sys/kernel/debug/tracing
when it feels the need.  This makes it rather tricky to count on and
script around at boot.

We've got to support older 4.9 kernels, so try both when asked.  This is
a setup time cost, not worth stressing about.

Change-Id: I4490616afb9e5bfb10326b8a96085e84b68d6447
Signed-off-by: Austin Schuh <austin.linux@gmail.com>