Fix FatalUnsetRealtimePriority
Errno was getting modified, so PCHECK wasn't printing the right thing.
Save it and restore it around the scheduler restore.

While I was debugging this with strace, I noticed that we weren't
switching to SCHED_OTHER properly, and that syscall was returning an
error (which was the errorno being displayed with PCHECK).

Change-Id: Iadd0c8725766318f1480ff2f50d51fa0a6a0516d
Signed-off-by: Austin Schuh <austin.linux@gmail.com>