Fix parsing for @pip target aliases
I tried adding `glog==0.1` to the requirements file. That ended up
generating an alias of `@pip//glog==0_1`. That's definitly not what we
want. We just want `@pip//glog`.

This patch fixes the issue by simplifying the parsing logic. Now all
comments and empty lines are handled by the same logic. A single regex
extracts the first package name on a line.

Signed-off-by: Philipp Schrader <philipp.schrader@gmail.com>
Change-Id: I9036ff668a74b06f3ec1b72e398e634eba5924f4