Fix dependencies in //frc971/control_loops/python
The rules_python toolchain is stricter about some of this stuff.
Switching to it revealed some misspecified dependencies. This patch
attempts to fix some of those dependencies.

Separately, quite a bit of code relies on Python's
hermeticity-breaking behaviour of adding the fully resolved path of
the file's directory to the import search path. That will be fixed in
a separate patch.

Signed-off-by: Philipp Schrader <philipp.schrader@gmail.com>
Change-Id: I9691cefec7e18834fa288d9e11b9817e87dac2c8