Make scouting test servers work with rules_cypress
The upcoming `rules_js` migration will change the way that JS-based
tests work. Instead of executing from the root directory, the tests
execute from the directory in which they're defined. That means that
the `//scouting:scouting_test` test will execute from the `scouting`
directory in the future.

To accommodate this, I changed `scouting_test_servers.py` to use the
runfiles API. This lets the script access its dependencies regardless
of which directory it's executing from.

I also included a change where the script can accept an optional file
descriptor. When set, the script will print `READY` on that file
descriptor when startup is complete. This lets the Cypress test
framework sync with the script easily. Otherwise, I kept running into
flaky behaviour.

Signed-off-by: Philipp Schrader <philipp.schrader@gmail.com>
Change-Id: I5fa973c97faf53ec4149d07987f7a5f5b0066285