Add package.json files for //scouting/www
I'm not sure if these will be useful yet, but as far as I can tell,
they don't hurt anything. They will be required for the new Angular
rules I'm importing for the rules_js migration.

It's possible we can get rid of these, but I'll figure that out in a
future patch.

Signed-off-by: Philipp Schrader <philipp.schrader@gmail.com>
Change-Id: Ice7b87d164bcbd557d93591d33370dcde5371835